Generate Predictions for New Data Using a Logistic Regression Model

This function generates predictions for new data based on an input logistic regression model fitted using function lr. The model is should include a list that contains an optimized set of coefficients (betas). Predictions are based on a logistic regression model using a specified threshold to classify outcomes.

predict_new(data, model, threshold = 0.5)

Arguments

data

A data frame containing the new data to make predictions on. The dataset should contain all predictors specified in the model.

model

A list representing the logistic regression model. It must contain a component named beta_optimized, which is a column matrix containing the model's coefficients, including the intercept.

threshold

A numeric value between 0 and 1. This value is used as the decision threshold for classifying outcomes as 0 or 1. The default value is 0.5.

Value

A data frame containing the original data along with an additional column named predicted, which holds the predicted class labels (0 or 1).

See also

Examples

# Example 1: Using the mtcars dataset with 'am' as the binomial dependent variable.

# For this example we will create a model object using function \code{lr} using the training data, and variable 'am' as the dependent variable.
data(mtcars)
model<-lr(am~hp+mpg+wt, data=mtcars, B=50, alpha=0.05)

# For this example we will generate random data for variables 'hp', 'mpg', and 'wt'.
n<-10
hp<-round(runif(n, min(mtcars$hp), max(mtcars$hp)),1)
mpg<-round(runif(n, min(mtcars$mpg), max(mtcars$mpg)),1)
wt<-round(runif(n, min(mtcars$wt), max(mtcars$wt)),1)

new_data<-data.frame(
  hp=hp,
  wt=wt,
  mpg=mpg)

#Now we use the function to predict the outcome
new_predictions<-predict_new(data=new_data, model=model, threshold=0.5)
head(new_predictions)
#>      hp  wt  mpg predicted
#> 1  54.7 1.8 31.1         1
#> 2 219.0 4.8 10.5         0
#> 3 175.8 2.6 19.6         1
#> 4 147.3 2.8 21.7         1
#> 5  72.8 2.5 20.3         0
#> 6 310.5 1.7 30.1         1

# Example 2: Predicting outcomes on a binary subset of the Iris dataset
data(iris)
iris_binary <- iris[iris$Species %in% c("setosa", "versicolor"), ]
iris_binary$Species <- factor(iris_binary$Species)

# Fit a logistic regression model using function \code{lr}
model <- lr(Species ~ Sepal.Length + Sepal.Width, data = iris_binary)

# For this example we will generate new random data for variables 'Sepal.Length' and 'Sepal.Width'.
n<-10
Sepal.Length<-round(runif(n, min(iris_binary$Sepal.Length), max(iris_binary$Sepal.Length)), 1)
Sepal.Width<-round(runif(n, min(iris_binary$Sepal.Width), max(iris_binary$Sepal.Width)), 1)
new_data<-data.frame(
  Sepal.Length=Sepal.Length,
  Sepal.Width=Sepal.Width)

# Predict outcomes on new data
predictions <- predict_new(data = new_data, model = model, threshold=0.5)
head(predictions)
#>   Sepal.Length Sepal.Width predicted
#> 1          4.3         2.8         0
#> 2          5.8         3.0         1
#> 3          4.9         2.9         0
#> 4          5.7         2.6         1
#> 5          5.1         2.5         1
#> 6          4.5         2.4         0