Train-Test Sampling Function

This function splits a data set into training and test sets based on a specified proportion of training data, while ensuring that the split maintains the class distribution of a binomial dependent variable.

train_test_sampling(
  data,
  dependent_var,
  train_prop = 0.75,
  return_data = FALSE,
  seed = NULL
)

Arguments

data

A data frame containing the data to be split. The data should have no missing values in the rows intended for modeling.

dependent_var

A character string or numeric value indicating the name or column index of the dependent (target) variable in `data`. The dependent variable must be binomial (i.e., it should have exactly two unique levels).

train_prop

A numeric value between 0 and 1 indicating the proportion of the data to include in the training set. The default value is 0.75.

return_data

A logical value indicating whether to return the actual training and test datasets (`TRUE`) or just the training indices (`FALSE`). Default is `FALSE`.

seed

An optional integer to set the random seed for reproducibility.

Value

Depending on the value of `return_data`, the function returns: - If `return_data = FALSE`: A vector of indices for the rows in the training set. - If `return_data = TRUE`: A list with two elements: - `train`: A data frame containing the training set. - `test`: A data frame containing the test set.

Details

The function first removes any incomplete cases from the data and then checks whether the dependent variable is binomial. It then proceeds to split the data into training and test sets based on the specified proportion (`train_prop`). If `return_data = TRUE`, the function returns a list containing the training and test datasets. Otherwise, it returns only the indices of the rows assigned to the training set.

See also

Examples

# Example 1: Using the mtcars dataset with 'am' as the binomial dependent variable.
# Split the data, using 75% of the data for training, and return the train and test datasets.

split_data <- train_test_sampling(mtcars, dependent_var = 'am', train_prop = 0.75, return_data = TRUE, seed = 123)

#To obtain the training data set can be found in an element called \code{train}
head(split_data$train)
#>                 mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Datsun 710     22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Volvo 142E     21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> Ford Pantera L 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Mazda RX4 Wag  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Toyota Corolla 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Maserati Bora  15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
dim(split_data$train)
#> [1] 24 11

#The test data set is stored in an element called \code{test}
head(split_data$test)
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
dim(split_data$test)
#> [1]  8 11

# Get the training indices only
train_indices <- train_test_sampling(mtcars, dependent_var = 'am', train_prop = 0.75, return_data = FALSE, seed = 123)

# Example 2: Splitting the Iris dataset (binary classification case)
data(iris)
iris_binary <- iris[iris$Species %in% c("setosa", "versicolor"), ]
iris_binary$Species <- factor(iris_binary$Species)

# Get indices for the training set
train_indices <- train_test_sampling(data = iris_binary, dependent_var = "Species", train_prop = 0.8)
head(train_indices)
#> [1] 12 15 32 42 45  7

# Split data into training and testing sets
split_data <- train_test_sampling(data = iris_binary, dependent_var = "Species", train_prop = 0.8, return_data = TRUE)
head(split_data$train)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 35          4.9         3.1          1.5         0.2  setosa
#> 40          5.1         3.4          1.5         0.2  setosa
#> 46          4.8         3.0          1.4         0.3  setosa
#> 30          4.7         3.2          1.6         0.2  setosa
#> 15          5.8         4.0          1.2         0.2  setosa
#> 24          5.1         3.3          1.7         0.5  setosa
head(split_data$test)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 10          4.9         3.1          1.5         0.1  setosa
#> 17          5.4         3.9          1.3         0.4  setosa
#> 18          5.1         3.5          1.4         0.3  setosa
#> 28          5.2         3.5          1.5         0.2  setosa
#> 31          4.8         3.1          1.6         0.2  setosa
#> 33          5.2         4.1          1.5         0.1  setosa